+++ covenant-signer/itest/bitcoindnodesetup.go
