+++ covenant-signer/itest/containers/config.go
@@ -10,7 +10,7 @@ type ImageConfig struct {
 //nolint:deadcode
 const (
 	dockerBitcoindRepository = "lncm/bitcoind"
-	dockerBitcoindVersionTag = "v26.0"
+	dockerBitcoindVersionTag = "v27.0"
 )
 
 // NewImageConfig returns ImageConfig needed for running e2e test.
