+++ covenant-signer/signerapp/hardcodedprivkeyretriever.go
